applications: testapp: move cert to bluetooth shell subsystem
Move cert.c to subsys/bluetooth/shell so it is available to all BLE
applications. Cert IDs are now read from Kconfig, set per board in
Kconfig.defconfig. Board selection menu removed as the board is
determined at compile time.
