+++ applications/testapp/CMakeLists.txt
@@ -18,5 +18,4 @@ project(test_application)
 
 target_sources(app PRIVATE
   src/main.c
-  src/cert.c
 )
