+++ applications/testapp/boards/alif_b1_dk_ab1c1f4m51820hh0_rtss_he.conf
@@ -34,3 +34,5 @@ CONFIG_MBEDTLS_HEAP_SIZE=30000
 # kernel log-level bt_att 4
 CONFIG_LOG_RUNTIME_FILTERING=y
 
+CONFIG_ALIF_CERT_SHELL=y
+
