+++ subsys/bluetooth/shell/CMakeLists.txt
@@ -10,3 +10,4 @@ zephyr_library()
 zephyr_library_sources(
   bt.c
 )
+zephyr_library_sources_ifdef(CONFIG_ALIF_CERT_SHELL cert.c)
